Made the Create Post page and updated the tab information.
