+++ yapp/app/dashboard/post-creation/page.tsx
@@ -1,0 +1,10 @@
+import ClientCreatePost from "./ClientCreatePost";
+
+export const metadata = {
+  title: "📝 Create Post",
+  description: "Write and share a new post on Yapp.",
+};
+
+export default function CreatePostPage() {
+  return <ClientCreatePost />;
+}
