+++ yapp/app/utils/mongodb.ts
@@ -1,0 +1,19 @@
+import { MongoClient } from "mongodb";
+
+const uri = process.env.MONGODB_URI!;
+const options = {};
+
+let client: MongoClient;
+let clientPromise: Promise<MongoClient>;
+
+if (!global._mongoClientPromise) {
+  client = new MongoClient(uri, options);
+  global._mongoClientPromise = client.connect();
+}
+clientPromise = global._mongoClientPromise;
+
+export async function connectToDatabase() {
+  const client = await clientPromise;
+  const db = client.db("yapp"); // Change "yapp" to your DB name
+  return { client, db };
+}
