+++ yapp/package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "firebase": "^11.4.0",
+    "firebase": "^11.6.0",
     "mongodb": "^6.15.0",
     "next": "^15.1.6",
     "react": "^19.0.0",
